requirements.txt
flask==2.2.4
asyncio
pandas
requests
colorama
typing
flask_session==0.4.0
sentry-sdk
parse
redis
flask-babel
pyyaml
# pysqlite3
flask-profiler
bs4
cryptography
pyjwt